Simplify (and fix) mulit-condition waiting
Previous version had a bug:
MultiConditon would each time add and remove external conditions from
its internal dict when it entered and exited `wait()`. This is only
correct if there is only one other thread waiting for a single conditon.
Instead, a counter is needed for each condition, so that each time a
unique thread waits for it, a counter goes up, and the condition is
removed from the dict only when that counter goes to 0.